requirements.txt
Pillow
future
futures
tqdm
numpy